BigInt for user_id

+++ src/db.py
@@ -1,6 +1,7 @@
 from sqlalchemy import Column, String, Date, Integer
 from sqlalchemy.orm import DeclarativeBase
 from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy import BigInteger
 import uuid
 
 
@@ -18,7 +19,7 @@ class User(Base):
         default=uuid.uuid4
     )
     first_name = Column(String())
-    user_id = Column(Integer(), unique=True, index=True, nullable=False)
+    user_id = Column(BigInteger(), unique=True, index=True, nullable=False)
     sub_expire_date = Column(Date())
